abstract out components, rename to AlgoliaSearch
