+++ src/views/docs-view/components/product-docs-search/product-docs-search.module.css
@@ -114,7 +114,6 @@
 
 .list {
   padding: 12px;
-  padding-right: 4px;
   list-style-type: none;
   margin: 0;
 }
@@ -155,20 +154,25 @@
   }
 }
 
+.enterIcon {
+  color: var(--token-color-foreground-faint);
+  display: none;
+  position: absolute;
+  right: 14px;
+  top: calc(50% - 8px);
+}
+
 .itemDescription {
   color: var(--token-color-foreground-faint);
   font-size: 14px;
+  display: -webkit-box;
+  -webkit-line-clamp: 2;
+  -webkit-box-orient: vertical;
+  overflow: hidden;
 
   & mark {
     color: var(--token-color-foreground-action);
     background: transparent;
-  }
-
-  & .enterIcon {
-    display: none;
-    position: absolute;
-    right: 14px;
-    top: calc(50% - 8px);
   }
 }
