+++ src/views/docs-view/components/product-docs-search/search-results-legend.tsx
@@ -1,0 +1,11 @@
+import Badge from 'components/badge'
+import s from './product-docs-search.module.css'
+
+export default function SearchResultsLegend() {
+  return (
+    <div className={s.legend}>
+      <Badge text="↑" /> <Badge text="↓" /> to navigate, <Badge text="ENTER" />{' '}
+      to select, <Badge text="ESC" /> to dismiss
+    </div>
+  )
+}
