+++ src/views/docs-view/components/product-docs-search/types.ts
@@ -1,0 +1,7 @@
+import { Hit } from '@algolia/client-search'
+
+export type AutocompleteItem = Hit<{
+  page_title: string
+  description: string
+  objectID: string
+}>
